feat: eventAPI 수정완료 (지역)
